# src/core/BrightstarDB.Server.AspNet/assets/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "BrightstarDB REST API",
        "description": "The BrightstarDB REST API provides access to the full functionality of the BrightstarDB\nserver. The server is modelled around a collection of \"stores\". Each store is an\nindependant RDF data store providing its own set of operations and its own SPARQL\nquery and update endpoints.\n",
        "license": {
            "name": "MIT License",
            "url": "http://opensource.org/licenses/MIT"
        }
    },
    "basePath": "/brightstar",
    "produces": [
        "application/json",
        "application/xml"
    ],
    "consumes": [
        "application/json",
        "application/xml"
    ],
    "paths": {
...
        "/{storeName}/update": {
            "get": {
                "tags": [
                    "sparql",
                    "update"
                ],
                "summary": "Executes a SPARQL Update operation against the store",
                "description": "",
                "operationId": "sparqlUpdate",
                "parameters": [
                    {
                        "name": "storeName",
                        "in": "path",
                        "description": "The name of the store",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "update",
                        "in": "query",
                        "description": "The SPARQL query string",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "using-graph-uri",
                        "in": "query",
                        "description": "The URI of a graph to add to the dataset",
                        "type": "string",
                        "required": false,
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "using-named-graph-uri",
                        "in": "query",
                        "description": "The URI of a named graph to add to the dataset",
                        "type": "string",
                        "format": "url",
                        "required": false,
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Bad Request. Either a required parameter was not provided;\nor the update could not be parsed.\n"
                    }
                }
            },
            "post": {
                "tags": [
                    "sparql",
                    "update"
                ],
                "summary": "Executes a SPARQL Update operation against the store.",
                "description": "NOTE: The Swagger spec only allows us to document one possible form of POST. However\nBrightstarDB fully supports all of the request options described in \n[SPARQL 1.1 Protocol](http://www.w3.org/TR/sparql11-protocol/)\n",
                "operationId": "sparqlUpdatePost",
                "parameters": [
                    {
                        "name": "storeName",
                        "in": "path",
                        "description": "The name of the store",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "update",
                        "in": "formData",
                        "description": "The SPARQL query string",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "using-graph-uri",
                        "in": "formData",
                        "description": "The URI of a graph to add to the dataset",
                        "type": "string",
                        "format": "url",
                        "required": false,
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "using-named-graph-uri",
                        "in": "formData",
                        "description": "The URI of a named graph to add to the dataset",
                        "type": "string",
                        "format": "url",
                        "required": false,
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Bad Request. Either a required parameter was not provided;\nor the update could not be parsed.\n"
                    }
                }
            }
        }
    },
...
    }